トップページの作り替え #13

+++ database/migrations/2023_08_06_112746_create_values_table.php
@@ -15,8 +15,8 @@ class CreateValuesTable extends Migration
     {
         Schema::create('values', function (Blueprint $table) {
             $table->id();
-            $table->double('high_value', 6, 3);
-            $table->double('row_value', 6, 3);
+            $table->decimal('high_value', 6, 3);
+            $table->decimal('row_value', 6, 3);
             $table->timestamps();
         });
     }
